+++ lib/utils/image_file_loader_io.dart
@@ -1,0 +1,7 @@
+import 'dart:io';
+import 'dart:typed_data';
+
+Future<Uint8List> readFileAsBytes(String path) async {
+  return await File(path).readAsBytes();
+}
+