+++ lib/utils/image_file_loader_stub.dart
@@ -1,0 +1,6 @@
+import 'dart:typed_data';
+
+Future<Uint8List> readFileAsBytes(String path) {
+  throw UnsupportedError('File loading is not supported on this platform.');
+}
+